sc: test ".uno:RemoveSheetView" to remove current sheet view
Also fix issues found with the test.

We need to broadcast that the table was inserted or it won't
properly be accounted for ViewData, which will cause an assert
to fail.

Reviewed-on: https://gerrit.libreoffice.org/c/core/+/189386
Tested-by: Jenkins
(cherry picked from commit e6cf2c90f14097a1c2ebbd013ccb06276054ad17)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190954
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>

+++ sc/source/ui/view/viewfun3.cxx
@@ -64,6 +64,7 @@
 #include <clipoptions.hxx>
 #include <gridwin.hxx>
 #include <SheetView.hxx>
+#include <uiitems.hxx>
 #include <com/sun/star/util/XCloneable.hpp>
 
 using namespace com::sun::star;
@@ -2070,6 +2071,9 @@ void ScViewFunc::MakeNewSheetView()
     SCTAB nSheetViewTab = nTab + 1;
     if (rDocument.CopyTab(nTab, nSheetViewTab))
     {
+        GetViewData().GetDocShell()->Broadcast(ScTablesHint(SC_TAB_INSERTED, nSheetViewTab));
+        SfxGetpApp()->Broadcast(SfxHint(SfxHintId::ScTablesChanged));
+
         // Add and register the created sheet view
         rDocument.SetSheetView(nSheetViewTab, true);
         sc::SheetViewID nSheetViewID = rDocument.CreateNewSheetView(nTab, nSheetViewTab);
@@ -2078,6 +2082,9 @@ void ScViewFunc::MakeNewSheetView()
         // Update
         GetViewData().SetTabNo(nSheetViewTab); // force add the sheet view tab
         GetViewData().SetTabNo(nTab); // then change back to the current tab
+
+        ScDocShell& rDocSh = *GetViewData().GetDocShell();
+        rDocSh.PostPaintGridAll();
         PaintExtras(); // update Tab Control
     }
 }
@@ -2088,18 +2095,19 @@ void ScViewFunc::RemoveCurrentSheetView()
     if (nSheetViewID == sc::DefaultSheetViewID)
         return;
 
-    SCTAB nTab = GetViewData().GetTabNumber();
     ScDocument& rDocument = GetViewData().GetDocument();
-    SCTAB nSheetViewTab = rDocument.GetSheetViewNumber(nTab, nSheetViewID);
-
+    SCTAB nTab = GetViewData().GetTabNumber();
     auto pSheetManager = rDocument.GetSheetViewManager(nTab);
+    if (!pSheetManager)
+        return;
+
+    SCTAB nSheetViewTab = rDocument.GetSheetViewNumber(nTab, nSheetViewID);
     pSheetManager->remove(nSheetViewID);
     GetViewData().SetSheetViewID(sc::DefaultSheetViewID);
     GetViewData().SetTabNo(nTab);
     GetViewData().GetDocFunc().DeleteTable(nSheetViewTab, true);
 
-    ScDocShell& rDocSh = *GetViewData().GetDocShell();
-    rDocSh.PostPaintGridAll();
+    GetViewData().GetDocShell()->PostPaintGridAll();
     PaintExtras();
 }
